Include quic native libraries for all platforms (for test module only). Extract Quic library version to versions.gradle. Move tests from testFixtures to test
