+++ libp2p/build.gradle.kts
@@ -10,7 +10,6 @@ dependencies {
     api("io.netty:netty-transport")
     implementation("io.netty:netty-handler")
     implementation("io.netty:netty-codec-http")
-    implementation("io.netty.incubator:netty-incubator-codec-native-quic:0.0.38.Final:linux-x86_64")
     implementation("io.netty:netty-transport-classes-epoll:4.1.90.Final")
 
     api("com.google.protobuf:protobuf-java")
@@ -22,6 +21,12 @@ dependencies {
     implementation("org.bouncycastle:bctls-jdk15on")
 
     testImplementation(project(":tools:schedulers"))
+
+    testImplementation("io.netty.incubator:netty-incubator-codec-native-quic::linux-x86_64")
+    testImplementation("io.netty.incubator:netty-incubator-codec-native-quic::linux-aarch_64")
+    testImplementation("io.netty.incubator:netty-incubator-codec-native-quic::osx-x86_64")
+    testImplementation("io.netty.incubator:netty-incubator-codec-native-quic::osx-aarch_64")
+    testImplementation("io.netty.incubator:netty-incubator-codec-native-quic::windows-x86_64")
 
     testFixturesApi("org.apache.logging.log4j:log4j-core")
     testFixturesImplementation(project(":tools:schedulers"))
